Update stud_util.py
Added a background task that send a week's worth of announcements periodically to the user's dms using the functionality that used to be in the automatic updates slash command. Slash command now can toggle and start the background task. Stores canvas session, user id, toggle choice, and api key in self to pass them from the slash command to the background task.